+++ cli/azd/internal/repository/testdata/empty/gitignore_existing.txt
