+++ cli/azd/internal/repository/testdata/empty/gitignore_existing_notrail.txt
