+++ cli/azd/internal/repository/testdata/empty/gitignore_with_env.txt
