+++ cli/azd/internal/repository/testdata/template-minimal/README.md.txt
