+++ cli/azd/internal/repository/testdata/template/azure.yaml.txt
